requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.25.0
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.2.14
langchain-community==0.2.12
openai==1.47.0
streamlit==1.38.0
requests==2.32.3
